+++ Assets/Prefabs/Gun.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e6c5bcd1ed2154de5b49e449a775af72
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
