--- Assets/Prefabs/PlayerConnectionObj.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b4d02b62ddd024af393f06b25abde066
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
