+++ Assets/Prefabs/PlayerUnit.prefab.meta
@@ -1,10 +1,7 @@
 fileFormatVersion: 2
-guid: eeb52913066634a43920cc0af5b50ef8
-timeCreated: 1515527935
-licenseType: Free
-NativeFormatImporter:
+guid: fc223806b39c44637b08f57b98901f86
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
