+++ Assets/Prefabs/PlayerUnitOld.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: eeb52913066634a43920cc0af5b50ef8
+timeCreated: 1515527935
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
