+++ Assets/Scripts/CountDownTimerUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a86838e6f176f46d9a847c8ba790d695
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
