--- Assets/Scripts/PlayerConnectionObj.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6bc9776bebf3147c9b1c6f70507670f4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
