+++ Assets/Scripts/PlayerUnit.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: 78b778c83e1a6ca4786f9a4b241f18c5
-timeCreated: 1515529325
-licenseType: Free
+guid: 52d4ef7807d7b4d18a0cf0ceac453b65
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
