+++ Assets/Scripts/PlayerUnitOld.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 78b778c83e1a6ca4786f9a4b241f18c5
+timeCreated: 1515529325
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
