--- Assets/Scripts/ThirdPersonClient.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 52d4ef7807d7b4d18a0cf0ceac453b65
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
