+++ ProjectSettings/TagManager.asset
@@ -41,3 +41,9 @@ TagManager:
   - name: Default
     uniqueID: 0
     locked: 0
+  - name: Canvas
+    uniqueID: 3419298505
+    locked: 0
+  - name: PlayerUnitPrefab
+    uniqueID: 579088849
+    locked: 0
